Convert to Search UI (#31)
* Replaced all code in this project by using the search-ui library. 

Since this project was the precursor to Search UI, it was very easy to swap out.

The only thing I couldn't carry over support for was the urlFieldTemplate configuration option because it would have required me writing a custom view for the Result component, which I didn't want to do.

* Fixed issue with react scripts build

React scripts was trying to build a bunch of "chunk" js files,
and we just needed a single file to include in the
website project. This fixes that issue.