Added Chrome 44 to get Arrow Functions
Without this, currently, support start at Chrome 48, which has
support for ES6 arrow functions, so this will no longer transpile
them to normal functions.

This breaks support with App Search when it hosts the built script.

By fixing this to at least Chrome 44, we are able to work around this
issue, as arrow functions will always be transpiled.